Use persistent identity high-water marks across authoring

+++ src/main/java/games/pixscape/studio/event/EventFlow.java
@@ -213,14 +213,14 @@ public final class EventFlow {
 
     public record FixtureSelectionChanged(
             int bodyEntityId,
-            long physicsShapeId,
+            int physicsShapeId,
             int sourceTag
     ) {
     }
 
     public record FixtureParametersChanged(
             int bodyEntityId,
-            long physicsShapeId,
+            int physicsShapeId,
             int sourceTag
     ) {
     }
