+++ src/main/java/games/pixscape/studio/ui/property/entityproperties/ConeLightProperties.java
@@ -299,7 +299,7 @@ public final class ConeLightProperties extends VisTable {
         currentEntityId = entityId;
 
         PixscapeIdentityComponent identity = ctx.mIdentity.getSafe(entityId, null);
-        long stableId = (identity != null && identity.stableId > 0L) ? identity.stableId : 0L;
+        int stableId = (identity != null && identity.stableId > 0) ? identity.stableId : 0;
         entityIdValueLabel.setText(String.valueOf(stableId));
         icon.setDrawable(ctx.iconResolver.iconForEntity(entityId));
