+++ src/main/java/games/pixscape/studio/ui/property/entityproperties/EntityProperties.java
@@ -206,7 +206,7 @@ public class EntityProperties extends VisTable {
         syncScenePhysicsEnabled();
 
         PixscapeIdentityComponent identity = ctx.mIdentity.getSafe(entityId, null);
-        long stableId = (identity != null && identity.stableId > 0L) ? identity.stableId : 0L;
+        int stableId = (identity != null && identity.stableId > 0) ? identity.stableId : 0;
         entityIdValueLabel.setText(String.valueOf(stableId));
         icon.setDrawable(ctx.iconResolver.iconForEntity(entityId));
